Make GP_CAMLIB(lumix) conditional instead of lumix.c code
Make GP_CAMLIB(lumix) macro invocation conditional instead
of the C code inside of the lumix camlib.

If either libcurl or libxml2 are missing, this will result
in compilation of lumix.c failing, as the macros GP_CAMLIB
and GP_CAMLIBS_DEFINE still need to be fixed to prevent the
lumix camlib from being built in that case (issue #467).